Move Visibility::Visit into its own file, use it for Visibility::Profile
